(1) Added two more tests for GTF feature column; (2) moved ParseGTF function to a different test class; (3) some refactoring.
